+++ src/TrackRowInfoVisQuantitativeBar.js
@@ -25,6 +25,8 @@ export const margin = 5;
  * @prop {object} fieldInfo The name and type of data field.
  * @prop {boolean} isLeft Is this view on the left side of the track?
  * @prop {string} titleSuffix The suffix of a title, information about sorting and filtering status.
+ * @prop {object} sortInfo The options for sorting rows of the field used in this track.
+ * @prop {object} filterInfo The options for filtering rows of the field used in this track.
  * @prop {function} onAddTrack The function to call upon a track insertion.
  * @prop {function} onSortRows The function to call upon a sort interaction.
  * @prop {function} onSearchRows The function to call upon a search interaction.
@@ -38,6 +40,8 @@ export default function TrackRowInfoVisQuantitativeBar(props) {
         fieldInfo,
         isLeft,
         titleSuffix,
+        sortInfo,
+        filterInfo,
         onAddTrack,
         onSortRows,
         onSearchRows,
@@ -50,6 +54,7 @@ export default function TrackRowInfoVisQuantitativeBar(props) {
     const canvasRef = useRef();
     const hiddenCanvasRef = useRef();
     const [isMouseHover, setIsMouseHover] = useState(null);
+    const [xExtent, setXExtent] = useState([null, null]);
 
     // Data, layouts and styles
     const { field } = fieldInfo;
@@ -261,6 +266,8 @@ export default function TrackRowInfoVisQuantitativeBar(props) {
         d3.select(canvas).on("mouseout", destroyTooltip);
         d3.select(div).on("mouseenter", () => setIsMouseHover(true));
         d3.select(div).on("mouseleave", () => setIsMouseHover(null));
+        
+        setXExtent(d3.extent(xScale.domain()));
 
         // Clean up.
         return () => {
@@ -316,10 +323,14 @@ export default function TrackRowInfoVisQuantitativeBar(props) {
                 fieldInfo={fieldInfo}
                 searchTop={top + axisHeight}
                 searchLeft={left}
+                sortAsceButtonHighlit={sortInfo && sortInfo.order === "ascending"}
+                sortDescButtonHighlit={sortInfo && sortInfo.order === "descending"}
+                filterButtonHighlit={filterInfo && filterInfo.length > 0}
                 onSortRows={onSortRows}
                 onSearchRows={onSearchRows}
                 onFilterRows={onFilterRows}
                 transformedRowInfo={transformedRowInfo}
+                valueExtent={xExtent}
             />
         </div>
     );